+++ packages/terra-table/src/subcomponents/ColumnHeaderCell.jsx
@@ -168,20 +168,20 @@ const ColumnHeaderCell = (props) => {
 
   const isGridContext = gridContext.role === GridConstants.GRID;
 
-  const columnHeaderFocusArea = useCallback(() => (columnHeaderCellButtonRef.current ? columnHeaderCellButtonRef.current : columnHeaderCellRef.current), []);
+  const columnHeaderFocusableElement = useCallback(() => (columnHeaderCellButtonRef.current ? columnHeaderCellButtonRef.current : columnHeaderCellRef.current), []);
 
   useEffect(() => {
     if (isActive) {
       if (isResizable && isResizeActive) {
         setResizeHandleActive(true);
       } else {
-        columnHeaderFocusArea().focus();
+        columnHeaderFocusableElement().focus();
         setResizeHandleActive(false);
       }
     } else {
       setResizeHandleActive(false);
     }
-  }, [columnHeaderFocusArea, isActive, isResizable, isResizeActive]);
+  }, [columnHeaderFocusableElement, isActive, isResizable, isResizeActive]);
 
   const onResizeHandleMouseDown = useCallback((event) => {
     event.stopPropagation();
@@ -192,9 +192,9 @@ const ColumnHeaderCell = (props) => {
 
   // Restore focus to column header after resize action is completed.
   const onResizeHandleMouseUp = useCallback(() => {
-    columnHeaderFocusArea().focus();
+    columnHeaderFocusableElement().focus();
     setResizeHandleActive(false);
-  }, [columnHeaderFocusArea]);
+  }, [columnHeaderFocusableElement]);
 
   // Handle column header selection via the mouse click.
   const handleMouseDown = (event) => {
@@ -216,7 +216,7 @@ const ColumnHeaderCell = (props) => {
         break;
       case KeyCode.KEY_LEFT:
         if (isResizable && isResizeHandleActive && isGridContext) {
-          columnHeaderFocusArea().focus();
+          columnHeaderFocusableElement().focus();
           setResizeHandleActive(false);
           event.stopPropagation();
           event.preventDefault();
@@ -259,12 +259,14 @@ const ColumnHeaderCell = (props) => {
     : null;
 
   // For tables, we want elements to be tabbable when selectable, but not anytime else.
-  let columnTabIndex = isSelectable ? 0 : undefined;
-
+  let buttonTabIndex = isSelectable ? 0 : undefined;
   if (isGridContext) {
     // For grids, we only want 1 tab stop. We then define the focus behavior in DataGrid.
-    columnTabIndex = isSelectable && displayName ? -1 : undefined;
+    buttonTabIndex = isSelectable && displayName ? -1 : undefined;
   }
+
+  // Determine if button element is required for column header
+  const hasButtonElement = isSelectable && displayName;
 
   return (
   /* eslint-disable react/forbid-dom-props */
@@ -275,7 +277,7 @@ const ColumnHeaderCell = (props) => {
         selectable: isSelectable,
         pinned: columnIndex < columnContext.pinnedColumnOffsets.length,
       })}
-      tabIndex={isGridContext ? -1 : undefined}
+      tabIndex={isGridContext && !hasButtonElement ? -1 : undefined}
       role="columnheader"
       scope="col"
       aria-sort={sortIndicator}
@@ -285,8 +287,8 @@ const ColumnHeaderCell = (props) => {
     >
       <div
         className={cx('header-container')}
-        {...isSelectable && displayName && { ref: columnHeaderCellButtonRef, role: 'button' }}
-        tabIndex={columnTabIndex}
+        {...hasButtonElement && { ref: columnHeaderCellButtonRef, role: 'button' }}
+        tabIndex={buttonTabIndex}
       >
         {errorIcon}
         <span>{displayName}</span>