+++ packages/terra-table/CHANGELOG.md
@@ -11,6 +11,7 @@
   * Fixed issue where column header cell was receiving tab focus instead of the button for table role.
   * Fixed column header left border style.
   * Fixed resize handle being communicated when non-header cell is focused.
+  * Fixed tab index issues in the column header cells.
   
 ## 5.1.1-alpha.0 - (October 25, 2023)